fix admin

+++ admin/model/account/deleteUser.php
@@ -1,12 +1,22 @@
 <?php
     include '../header/header.php';
     $db = new DataBase();
-    $key = $_GET['id'];
-    $sql = "DELETE FROM user WHERE ID='$key'";
+    $key = $_POST['id'];
+    $sql = "SELECT type FROM user WHERE ID = '$key'";
     $result = $db->send($sql);
-    if ($result){
-        echo "delete account successfully $key";
+    $type = 1;
+    while($rows = $result->fetch_array()){
+        $type = $rows[0];
+    }
+    if($type != 0){
+        $sql = "DELETE FROM user WHERE ID='$key'";
+        $result = $db->send($sql);
+        if ($result){
+            echo "delete account successfully $key";
+        }else{
+            echo "SQL ERR";
+        }
     }else{
-        echo "Can't delete account";
+        echo "Can't delete admin account";
     }
 ?>
